consolidate forms
