+++ src/components/forms/components/field-wrapper/styles.scss
@@ -1,7 +1,6 @@
 @import '~styles/settings.scss';
 
 .c-form-field {
-  font-size: 14px;
   margin-bottom: 25px;
   display: flex;
   flex-direction: column;
