+++ src/components/forms/components/input/styles.scss
@@ -2,17 +2,10 @@
 
 .c-form-input {
   border-radius: 4px;
-  background-color: transparent;
   border: solid 1px $border;
-  outline: none;
-  color: $slate;
   height: 40px;
   padding: 0 12px;
   width: 100%;
-
-  &::placeholder {
-    color: #a8a8a8;
-  }
 
   &.textarea {
     height: 110px;
