+++ src/components/forms/components/select/styles.scss
@@ -11,7 +11,7 @@
   padding: 0 12px;
   font-size: 14px;
   color: $slate;
-  font-family: $font-family;
+  font-family: inherit;
   cursor: pointer;
 
   &.multiple {
