+++ src/components/forms/contact-us/styles.scss
@@ -3,11 +3,9 @@
 .c-contact-form {
   .subtitle {
     margin-bottom: 30px;
-    font-size: 16px;
   }
 
   h4 {
-    font-size: 16px;
     margin-bottom: 5px;
   }
 
@@ -33,7 +31,6 @@
     max-width: 500px;
 
     h3 {
-      font-size: 20px;
       margin-bottom: 30px;
       text-transform: none;
       opacity: 1;
